pyproject.toml
[project]
name = "flop"
version = "1.0.0"
description = "Streamlined OAuth authentication library for Flask microservices with automatic provider setup and optional global route protection"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
authors = [
    {name = "Flop Contributors"},
]
keywords = ["flask", "oauth", "authentication", "microservices", "google", "github"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Flask",
    "Topic :: Internet :: WWW/HTTP :: Session",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "authlib>=1.6.0",
    "flask>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.25.0",
    "werkzeug>=2.0.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
]
production = [
    "gunicorn>=21.0.0",
]

[project.urls]
Homepage = "https://github.com/your-username/flop"
Repository = "https://github.com/your-username/flop"
Issues = "https://github.com/your-username/flop/issues"
Changelog = "https://github.com/your-username/flop/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flop"]

[tool.pytest.ini_options]
testpaths = ["test_auth.py"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
